requirements.txt
argparse==1.4.0
pybind11==2.2.3
yafasttext==0.8.22.2
gensim==3.6.0
numpy==1.14.5
ujson==1.35
mxnet==1.3.0
gluonnlp==0.4.0
marisa_trie==0.7.5
tqdm==4.26.0